+++ Libraries/LibWeb/Fetch/Fetching/FetchedDataReceiver.cpp
@@ -11,7 +11,7 @@
 #include <LibWeb/Fetch/Infrastructure/Task.h>
 #include <LibWeb/HTML/Scripting/ExceptionReporter.h>
 #include <LibWeb/HTML/Scripting/TemporaryExecutionContext.h>
-#include <LibWeb/Streams/AbstractOperations.h>
+#include <LibWeb/Streams/ReadableStream.h>
 #include <LibWeb/WebIDL/Promise.h>
 
 namespace Web::Fetch::Fetching {
