+++ Libraries/LibWeb/Streams/ReadableByteStreamController.cpp
@@ -13,6 +13,7 @@
 #include <LibWeb/Streams/ReadableStream.h>
 #include <LibWeb/Streams/ReadableStreamBYOBRequest.h>
 #include <LibWeb/Streams/ReadableStreamDefaultReader.h>
+#include <LibWeb/Streams/ReadableStreamOperations.h>
 #include <LibWeb/WebIDL/Buffers.h>
 
 namespace Web::Streams {
