+++ Libraries/LibWeb/Streams/ReadableByteStreamController.h
@@ -10,7 +10,7 @@
 #include <AK/Optional.h>
 #include <AK/SinglyLinkedList.h>
 #include <LibWeb/Bindings/PlatformObject.h>
-#include <LibWeb/Streams/AbstractOperations.h>
+#include <LibWeb/Streams/Algorithms.h>
 
 namespace Web::Streams {
