+++ Libraries/LibWeb/Streams/ReadableStreamBYOBRequest.cpp
@@ -10,6 +10,7 @@
 #include <LibWeb/Bindings/ReadableStreamBYOBRequestPrototype.h>
 #include <LibWeb/Streams/ReadableByteStreamController.h>
 #include <LibWeb/Streams/ReadableStreamBYOBRequest.h>
+#include <LibWeb/Streams/ReadableStreamOperations.h>
 #include <LibWeb/WebIDL/Buffers.h>
 
 namespace Web::Streams {
