+++ Libraries/LibWeb/Streams/Transformer.cpp
@@ -5,7 +5,6 @@
  */
 
 #include <LibJS/Runtime/VM.h>
-#include <LibWeb/Streams/AbstractOperations.h>
 #include <LibWeb/Streams/Transformer.h>
 #include <LibWeb/WebIDL/CallbackType.h>
